Fix signed/unsigned mismatch warnings

+++ WinQuake/r_draw.cpp
@@ -458,7 +458,7 @@ void R_RenderFace (msurface_t *fa, int clipflags)
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED)
 				{
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount)
+						static_cast<unsigned int>( r_framecount ) )
 					{
 						r_lastvertvalid = false;
 						continue;
@@ -502,7 +502,7 @@ void R_RenderFace (msurface_t *fa, int clipflags)
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED)
 				{
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount)
+						static_cast<unsigned int>( r_framecount ) )
 					{
 						r_lastvertvalid = false;
 						continue;
